Module Ez_search.V1

module EzSearch : sig ... end

This module implements full-text search with regexps in a set of files. Two steps are required: in the first step, a database is generated from all the files; in the second step, searches are performed in the database.