Module V1.EzSearch

module TYPES : sig ... end

val index_directory : db_dir:string -> string -> unit

index_directory ~db_dir DIRECTORY index all files in DIRECTORY, and store the index in db_dir. Every top-directory in DIRECTORY is considered as a file_entry name, and file_name are within top-directories.

val load_db : db_dir:string -> ?use_mapfile:bool -> unit -> TYPES.db

load_db ~db_dir loads the database in memory. Note: in the next version, we should aim at using mmap.

val count_lines_total : db:TYPES.db -> int

count_lines_total ~db counts the number of '\n' in the database.

val search : db:TYPES.db -> f:(TYPES.occurrence -> bool) -> ?pos:int -> ?last:TYPES.occurrence -> Re.Str.regexp -> unit

search ~db ~f ?pos ?last regexp searches a regexp in the database, starting either from pos, from after the last occurrence last, or from the beginning. Calls f for every occurrence found. f returns a boolean, that should be true if the search should continue after, or false if the search should terminate immediately.

val occurrence_line : db:TYPES.db -> TYPES.occurrence -> int

occurrence_line ~db occ returns the line number in the file.

val occurrence_context : db:TYPES.db -> line:int -> TYPES.occurrence -> max:int -> TYPES.context

occurrence_context ~db ~line occ ~max returns the context of the occurrence of in the file. The line number of the occurrence, as provided by occurrence_line should be provided. The parameter max controls how many lines should be returned before and after the occurrence.

val time : string -> ('a -> 'b) -> 'a -> 'b

time msg f x prints the time spent executing f x.